Type the `useEvent` function
